+++ CHANGELOG.md
@@ -27,6 +27,7 @@ high state of flux, you're at risk of it changing without notice.
 - **Polish**
   - use `t.type` instead of `t.interface` (@gcanti)
   - use `t.record` instead of `t.dictionary`(@gcanti)
+  - recursive type output (@gcanti)
 
 # 0.2.2